# api/openapi/cbizdocsmanager-v1.yaml
openapi: 3.0.3
info:
  title: CBizDocsManager API
  version: 1.0.0
  description: 公务单据管理系统后端 API v1 契约。
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/groups/invitations:
    post:
      tags: [Group]
      operationId: createInvitation
      summary: 组主账号创建子账号邀请码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: 邀请码创建成功；明文邀请码仅在本响应返回
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInvitationApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无组管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Group]
      operationId: listInvitations
      summary: 组主账号分页查询邀请码列表
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/InvitationDisplayStatus'
      responses:
        '200':
          description: 查询成功；列表永不返回邀请码明文与密文
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationPageApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无组管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/groups/invitations/{invitation_id}/secret:
    post:
      tags: [Group]
      operationId: revealInvitation
      summary: 组主账号查看仍有效的邀请码明文
      description: 仅 active 且未过期的邀请码可查看；响应带 no-store，客户端展示后必须立即清除本地副本。
      parameters:
        - $ref: '#/components/parameters/InvitationIdPath'
      responses:
        '200':
          description: 查看成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
            Cache-Control:
              description: 固定为 no-store，禁止代理与浏览器缓存明文邀请码
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationSecretApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无组管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 邀请码已使用、已撤销或已过期，不再可查看（INVITATION_NOT_REVEALABLE）
          x-error-codes: [INVITATION_NOT_REVEALABLE]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          description: 密钥缺失或密文校验失败（INTERNAL_ERROR、INVITATION_DECRYPT_FAILED）
          x-error-codes: [INTERNAL_ERROR, INVITATION_DECRYPT_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
  /api/v1/groups/invitations/{invitation_id}/revoke:
    post:
      tags: [Group]
      operationId: revokeInvitation
      summary: 组主账号撤销仍有效的邀请码
      parameters:
        - $ref: '#/components/parameters/InvitationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeInvitationRequest'
      responses:
        '200':
          description: 撤销成功
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationSummaryApiResponse'
        '400':
          description: 请求参数错误（VALIDATION_FAILED）
          x-error-codes: [VALIDATION_FAILED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '401':
          $ref: '#/components/responses/TokenExpired'
        '403':
          description: 无组管理权限或必须先修改初始密码（FORBIDDEN、AUTH_PASSWORD_CHANGE_REQUIRED）
          x-error-codes: [FORBIDDEN, AUTH_PASSWORD_CHANGE_REQUIRED]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: 邀请码不存在（INVITATION_NOT_FOUND）
          x-error-codes: [INVITATION_NOT_FOUND]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '409':
          description: 邀请码不可撤销或已被其他请求修改（INVITATION_NOT_REVOKABLE、RESOURCE_VERSION_CONFLICT）
          x-error-codes: [INVITATION_NOT_REVOKABLE, RESOURCE_VERSION_CONFLICT]
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestID:
      description: 本次请求的唯一追踪 ID
      schema:
        type: string
  parameters:
...
    InvitationIdPath:
      name: invitation_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    TokenExpired:
      description: 访问令牌无效或已过期（AUTH_TOKEN_EXPIRED）
      x-error-codes: [AUTH_TOKEN_EXPIRED]
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
    InternalError:
      description: 服务内部错误（INTERNAL_ERROR）
      x-error-codes: [INTERNAL_ERROR]
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - AUTH_INVALID_CREDENTIALS
        - AUTH_TOKEN_EXPIRED
        - AUTH_REFRESH_INVALID
        - AUTH_PASSWORD_CHANGE_REQUIRED
        - USER_USERNAME_EXISTS
        - INVITATION_INVALID
        - INVITATION_EXPIRED
        - INVITATION_USED
        - INVITATION_NOT_FOUND
        - INVITATION_NOT_REVEALABLE
        - INVITATION_NOT_REVOKABLE
        - INVITATION_DECRYPT_FAILED
        - GROUP_NAME_EXISTS
        - GROUP_NOT_FOUND
        - OWNER_TARGET_INVALID
        - MEMBER_NOT_FOUND
        - MEMBER_OWNER_PROTECTED
        - MEMBER_SELF_OPERATION_FORBIDDEN
        - PERMISSION_CODE_INVALID
        - DICTIONARY_NOT_FOUND
        - DICTIONARY_NAME_EXISTS
        - DICTIONARY_PARENT_INVALID
        - RESOURCE_VERSION_CONFLICT
        - CSRF_INVALID
        - ORIGIN_FORBIDDEN
        - FORBIDDEN
        - INTERNAL_ERROR
    FieldError:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ApiResponse:
      type: object
      required: [code, message, data, request_id]
      properties:
        code:
          type: string
          description: 成功标识或稳定错误码
        message:
          type: string
        data:
          type: object
          nullable: true
          additionalProperties: true
        request_id:
          type: string
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    ErrorApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required: [code, message, data, request_id]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            data:
              type: object
              nullable: true
            field_errors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
...
    CreateInvitationRequest:
      type: object
      additionalProperties: false
      properties:
        expires_in_days:
          type: integer
          minimum: 1
          maximum: 30
          default: 7
...
    GroupSummary:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
...
    InvitationCreatedData:
      type: object
      additionalProperties: false
      required: [invitation_id, invitation_code, group, expires_at]
      properties:
        invitation_id:
          type: integer
          format: int64
          minimum: 1
        invitation_code:
          type: string
          description: 邀请码明文；后续查询不得再次返回
        group:
          $ref: '#/components/schemas/GroupSummary'
        expires_at:
          type: string
          format: date-time
    CreateInvitationApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/InvitationCreatedData'
...
    InvitationDisplayStatus:
      type: string
      description: 展示状态；active 过期后按 expired 投影，不改变库中状态。
      enum: [active, expired, used, revoked]
    InvitationSummaryData:
      type: object
      additionalProperties: false
      required: [invitation_id, status, expires_at, created_at, version]
      properties:
        invitation_id:
          type: integer
          format: int64
          minimum: 1
        status:
          $ref: '#/components/schemas/InvitationDisplayStatus'
        expires_at:
          type: string
          format: date-time
        used_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          minimum: 1
    InvitationSummaryApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/InvitationSummaryData'
    InvitationPageData:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvitationSummaryData'
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total:
          type: integer
          format: int64
          minimum: 0
    InvitationPageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/InvitationPageData'
    InvitationSecretData:
      type: object
      additionalProperties: false
      required: [invitation_id, invitation_code, expires_at]
      properties:
        invitation_id:
          type: integer
          format: int64
          minimum: 1
        invitation_code:
          type: string
          description: 邀请码明文；响应固定带 no-store，客户端用完必须立即清除。
        expires_at:
          type: string
          format: date-time
    InvitationSecretApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/InvitationSecretData'
    RevokeInvitationRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version:
          type: integer
          format: int64
          minimum: 1
          description: 客户端持有的邀请码版本号，用于乐观锁校验。
...